Add user authentication and authorization
